parser: Avoid reading after the end of shunting yard stack
We would previously do this:

- Erase items from the buffer
- Read the items we just erased

This is a mistake because the vector might have resized itself after the
erase call. This would cause assertion errors with the GNU STL. We now
instead read the offsets we want to read from before performing the
erase.

There may be other occurrences of this, but I've only hit these cases
when compiling Xonotic.